# Conflicts:
#	server/service/system/sys_autocode_history.go
#	web/src/view/superAdmin/authority/components/apis.vue

+++ server/service/system/sys_autocode_history.go
@@ -73,26 +73,28 @@ func (autoCodeHistoryService *AutoCodeHistoryService) RollBack(info *systemReq.R
 		return err
 	}
 	// 清除API表
-
-	ids := request.IdsReq{}
-	idsStr := strings.Split(md.ApiIDs, ";")
-	for i := range idsStr[0 : len(idsStr)-1] {
-		id, err := strconv.Atoi(idsStr[i])
-		if err != nil {
-			return err
+	var err error
+	if info.DeleteApi {
+		ids := request.IdsReq{}
+		idsStr := strings.Split(md.ApiIDs, ";")
+		for i := range idsStr[0 : len(idsStr)-1] {
+			id, err := strconv.Atoi(idsStr[i])
+			if err != nil {
+				return err
+			}
+			ids.Ids = append(ids.Ids, id)
 		}
-		ids.Ids = append(ids.Ids, id)
+		err = ApiServiceApp.DeleteApisByIds(ids)
+		if err != nil {
+			global.GVA_LOG.Error("ClearTag DeleteApiByIds:", zap.Error(err))
+		}
 	}
-	err := ApiServiceApp.DeleteApisByIds(ids)
-
-	if err != nil {
-		global.GVA_LOG.Error("ClearTag DeleteApiByIds:", zap.Error(err))
-	}
-
-	err = BaseMenuServiceApp.DeleteBaseMenu(int(md.MenuID))
-
-	if err != nil {
-		global.GVA_LOG.Error("ClearTag DeleteBaseMenu:", zap.Error(err))
+	// 清除菜单表
+	if info.DeleteMenu {
+		err = BaseMenuServiceApp.DeleteBaseMenu(int(md.MenuID))
+		if err != nil {
+			global.GVA_LOG.Error("ClearTag DeleteBaseMenu:", zap.Error(err))
+		}
 	}
 
 	// 删除表
